Remove duplicate skill model definitions
